Re-organized FITS and HDF5 readers
